provisory refactoring for passport and express
